Laba14 fix

+++ Laba14Var9/mat.h
@@ -1,6 +1,6 @@
 #ifndef MAT_H
 #define MAT_H
-void fillmat(int** mat, int m, int n, int k);
-void printmat(int** mat, int n, int m);
-void maxmat(int** mat, int n, int m, int* mxn, int* mxm, int* max);
+void fillmat(double** mat, int m, int n, int k);
+void printmat(double** mat, int n, int m);
+void maxmat(double** mat, int n, int m, int* mxn, int* mxm, double* max);
 #endif // !MAT_H
